Implement and expose a instrumentHTTP function
This commit implements the instrumentHTTP function. This function can
be used in scripts where the http default module is imported, in order
to automatically add tracing context to the requests it emits; without
any changes from the user to their script.

This is done by wrapping its delete, get, head, options, patch, post,
put and request methods so that it essentially fills the request's
params argument on the fly with the tracing headers generated by the
selected propagator. It will also ensure that the HTTP call emits
the generated trace id as an output metadata under the trace_id key.

Note that this commit does not implement support for the batch method
 for technical reasons justifying more research and an implementation
in a separate context/commit.