Define a tracing Client
This commit defines a Client type which exposes an API similar to the
HTTP module, and allows users to perform HTTP requests embedding a
trace-context, and attaching their trace_id as an output metadata to
their HTTP-related samples.

It uses the HTTP module's request function under the hood, and wraps it
with some "hook" logic to ensure the expected headers are present, and
that the expected output metadata are emitted.

The tracing client exposes an option-set allowing to use the three
available propagation format. Other options such as sampling, and
baggage are present for forward-compatibility, but will be implemented
at a later point in time.