Define a instrumentHTTP function
This commit implements the instrumentHTTP function. This function allows
to automatically wrap any further HTTP requests made with k6's HTTP
module with a trace context, and have those requests emit their trace id
as part of the output samples metadata.

Once imported and called with the same options-set as the Client, any
further http.* methods calls will transparently be instrumented, without
any modification needed in the user script.

This function relies heavily on the module holding a default instance of
a Client, and overwrites the imported http module's operation with calls
to Client's methods under the hood.